+++ InitiativeTracker.Domain/Concrete/EFCharacterRepository.cs
@@ -14,6 +14,21 @@ namespace InitiativeTracker.Domain.Concrete
     {
         private EFdbContext context = new EFdbContext();
 
+        private IEnumerable<CharacterGroup> groups
+        {
+            get { return context.CharacterGroup; }
+        }
+
+        private IEnumerable<EncounterCharacter> encounterCharacters
+        {
+            get { return context.ECL; }
+        }
+
+        private IEnumerable<Encounter> encounters
+        {
+            get { return context.Encounters; }
+        }
+
         public IEnumerable<Character> items
         {
             get { return context.Characters; }
@@ -45,14 +60,21 @@ namespace InitiativeTracker.Domain.Concrete
             return item;
         }
 
-        public void Remove(int id)
+        public string Remove(int id)
         {
             Character item = Get(id);
-            if (item != null)
+            if (item != null && encounterCharacters.Where(c => c.Character_ID == id).FirstOrDefault() == null)
             {
                 context.Characters.Remove(item);
                 context.SaveChanges();
+                return "Success";
+            } else
+            {
+                int encounter_ID = encounterCharacters.Where(c => c.Character_ID == id).FirstOrDefault().Encounter_ID;
+                string encounterName = encounters.Where(e => e.Encounter_ID == encounter_ID).FirstOrDefault().Name;
+                return string.Format("Character could not be deleted because they are part of the encounter named: {0}", encounterName);
             }
+
         }
     }
 }
